requirements.txt
fastapi
uvicorn[standard]
tensorflow
Pillow
httpx
numpy
python-multipart
pydantic-settings
pytest
